depthanything_v2_test/requirements.txt
# 핵심 딥러닝 프레임워크
torch>=2.0.0
torchvision>=0.15.0

# 웹 서버
flask>=2.3.0
gunicorn>=21.0.0

# 이미지 처리
opencv-python>=4.8.0
Pillow>=10.0.0

# 수치 계산
numpy>=1.24.0

# 시각화
matplotlib>=3.7.0

# 모델 관련
timm>=0.9.0
huggingface_hub>=0.16.0

# 추가 유틸리티
requests>=2.31.0
tqdm>=4.65.0 